fix: corrige estabilidad del runtime canonico y las puertas de CI

+++ include/finance.h
@@ -122,8 +122,8 @@ typedef enum {
     MILENA_DAY_COUNT_ACTUAL_360
 } MilenaDayCount;
 
-MilenaStatus milena_date_init(MilenaDate *out, int32_t year, uint8_t month,
-                              uint8_t day, MilenaError *error);
+MilenaStatus milena_date_init(MilenaDate *out, int32_t year, unsigned month,
+                              unsigned day, MilenaError *error);
 MilenaStatus milena_date_compare(const MilenaDate *left, const MilenaDate *right,
                                  int *result, MilenaError *error);
 MilenaStatus milena_date_days_between(const MilenaDate *start,
